// src/domains/service-clients/user/proto/generated/user/types/stats_types.ts
export interface InstructorsStats {
  total: number;
  active: number;
  inactive: number;
  blocked: number;
  newThisMonth: number;
}
// ...
export const InstructorsStats: MessageFns<InstructorsStats> = {
  encode(message: InstructorsStats, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.total !== 0) {
      writer.uint32(8).int32(message.total);
    }
    if (message.active !== 0) {
      writer.uint32(16).int32(message.active);
    }
    if (message.inactive !== 0) {
      writer.uint32(24).int32(message.inactive);
    }
    if (message.blocked !== 0) {
      writer.uint32(32).int32(message.blocked);
    }
    if (message.newThisMonth !== 0) {
      writer.uint32(40).int32(message.newThisMonth);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): InstructorsStats {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseInstructorsStats();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.total = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.active = reader.int32();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.inactive = reader.int32();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.blocked = reader.int32();
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.newThisMonth = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): InstructorsStats {
    return {
      total: isSet(object.total) ? globalThis.Number(object.total) : 0,
      active: isSet(object.active) ? globalThis.Number(object.active) : 0,
      inactive: isSet(object.inactive) ? globalThis.Number(object.inactive) : 0,
      blocked: isSet(object.blocked) ? globalThis.Number(object.blocked) : 0,
      newThisMonth: isSet(object.newThisMonth)
        ? globalThis.Number(object.newThisMonth)
        : isSet(object.new_this_month)
        ? globalThis.Number(object.new_this_month)
        : 0,
    };
  },

  toJSON(message: InstructorsStats): unknown {
    const obj: any = {};
    if (message.total !== 0) {
      obj.total = Math.round(message.total);
    }
    if (message.active !== 0) {
      obj.active = Math.round(message.active);
    }
    if (message.inactive !== 0) {
      obj.inactive = Math.round(message.inactive);
    }
    if (message.blocked !== 0) {
      obj.blocked = Math.round(message.blocked);
    }
    if (message.newThisMonth !== 0) {
      obj.newThisMonth = Math.round(message.newThisMonth);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<InstructorsStats>, I>>(base?: I): InstructorsStats {
    return InstructorsStats.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<InstructorsStats>, I>>(object: I): InstructorsStats {
    const message = createBaseInstructorsStats();
    message.total = object.total ?? 0;
    message.active = object.active ?? 0;
    message.inactive = object.inactive ?? 0;
    message.blocked = object.blocked ?? 0;
    message.newThisMonth = object.newThisMonth ?? 0;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}